% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{aggregate_pips}
\alias{aggregate_pips}
\title{Aggregate Posterior Inclusion Probabilities (PIPs)}
\usage{
aggregate_pips(alpha)
}
\arguments{
\item{alpha}{L x p matrix of per-effect inclusion probabilities.}
}
\value{
Length-p vector of marginal variable inclusion probabilities.
}
\description{
Aggregate Posterior Inclusion Probabilities (PIPs)
}
\keyword{internal}
