% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SER_ss.R
\name{BF_ss}
\alias{BF_ss}
\title{Bayesian Factor with Summary Statistics (log-space)}
\usage{
BF_ss(Xty, dXtX, sigma_2, sigma_0_2, mu_0)
}
\arguments{
\item{Xty}{p-vector of marginal associations X^T y.}

\item{dXtX}{p-vector = diag(X^T X).}

\item{sigma_2}{Residual variance (scalar).}

\item{sigma_0_2}{Prior effect variance (length p vector for the effect).}

\item{mu_0}{Prior effect mean (length p vector for the effect).}
}
\value{
Numeric p-vector of log Bayes factors for inclusion vs null.
}
\description{
Bayesian Factor with Summary Statistics (log-space)
}
