% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{build_temperature_schedule}
\alias{build_temperature_schedule}
\title{Build annealing temperature schedule (internal)}
\usage{
build_temperature_schedule(
  max_iter,
  start_T,
  schedule_type = c("geometric", "linear"),
  burn_in = 0
)
}
\arguments{
\item{max_iter}{Integer, number of iterations.}

\item{start_T}{Starting temperature (>= 1).}

\item{schedule_type}{One of "geometric" or "linear".}

\item{burn_in}{Number of initial iterations to anneal; after this, T=1.}
}
\value{
Numeric vector of length \code{max_iter} with temperatures per iteration.
}
\description{
Build annealing temperature schedule (internal)
}
\keyword{internal}
