% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_computation.R
\name{compute_Xb}
\alias{compute_Xb}
\title{Compute X \%*\% b using precomputed scaling}
\usage{
compute_Xb(X, b)
}
\arguments{
\item{X}{n x p (unstandardized) matrix with attributes \code{"scaled:center"},
\code{"scaled:scale"} and \code{"d"} set (see \code{initialize_X}).}

\item{b}{Length-p coefficient vector.}
}
\value{
Length-n numeric vector equal to X \%*\% b on the original scale.
}
\description{
Compute X \%*\% b using precomputed scaling
}
