% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_computation.R
\name{compute_Xty}
\alias{compute_Xty}
\title{Compute t(X) \%*\% y using precomputed scaling}
\usage{
compute_Xty(X, y)
}
\arguments{
\item{X}{n x p (unstandardized) matrix with attributes as in \code{compute_Xb}.}

\item{y}{Length-n response vector.}
}
\value{
Length-p numeric vector equal to t(X) \%*\% y on the standardized scale.
}
\description{
Compute t(X) \%*\% y using precomputed scaling
}
