% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERSS_ss.R
\name{ERSS_ss}
\alias{ERSS_ss}
\title{Expected Residual Sum of Squares (summary stats)}
\usage{
ERSS_ss(XtX, Xty, yty, b_hat, b_2_hat)
}
\arguments{
\item{XtX}{p x p cross-product matrix X^T X (on standardized scale).}

\item{Xty}{p-vector X^T y (on standardized scale).}

\item{yty}{Scalar y^T y.}

\item{b_hat}{L x p matrix of posterior mean effects.}

\item{b_2_hat}{L x p matrix of posterior second moments \eqn{E[b^2]}.}
}
\value{
Numeric scalar ERSS = \eqn{E[||y - X sum_l b_l||^2]}.
}
\description{
Expected Residual Sum of Squares (summary stats)
}
