% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize_model_fit}
\alias{initialize_model_fit}
\title{Initialize model fit traces (internal)}
\usage{
initialize_model_fit(
  max_iter,
  var_y,
  p,
  sigma2_init = NULL,
  elbo_init = NULL,
  residuals_init = NULL
)
}
\arguments{
\item{max_iter}{Maximum number of iterations.}

\item{var_y}{Variance of y (to initialize \code{sigma_2[1]}).}

\item{p}{Number of predictors (used to size residuals placeholder).}
}
\value{
List with fields \code{sigma_2}, \code{elbo}, \code{alpha_diff}, and
\code{full_residuals}.
}
\description{
Initialize model fit traces (internal)
}
\keyword{internal}
