% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize_X}
\alias{initialize_X}
\title{Initialize the X matrix}
\usage{
initialize_X(X)
}
\arguments{
\item{X}{(n x p) Data matrix}
}
\value{
X - a standardized matrix with column means (cm), column std.
deviations (csd), rowSums(X^2) (d) saved as pre-computed attributes.
}
\description{
Initialize the X matrix
}
