% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize_XtX}
\alias{initialize_XtX}
\title{Initialize the XtX matrix (summary stats)}
\usage{
initialize_XtX(XtX, n)
}
\arguments{
\item{XtX}{p x p cross-product matrix X^T X (original scale).}

\item{n}{Sample size.}
}
\value{
Standardized \code{XtX} with attributes \code{"d" = diag(XtX)} and
`"scaled:scale" = column standard deviations of X.
}
\description{
Initialize the XtX matrix (summary stats)
}
\keyword{internal}
