% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SER.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Log-Sum-Exp Utility}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Scalar log(sum(exp(x))) computed stably.
}
\description{
Log-Sum-Exp Utility
}
