% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_priors_ss.R
\name{neg_log_lik_ss}
\alias{neg_log_lik_ss}
\title{Negative log-likelihood under SER with summary stats (internal)}
\usage{
neg_log_lik_ss(priors, Xty, dXtX, sigma_2)
}
\arguments{
\item{priors}{One-row priors for a single effect.}

\item{Xty}{p-vector X^T y (standardized scale).}

\item{dXtX}{p-vector = diag(X^T X).}

\item{sigma_2}{Residual variance.}
}
\value{
Scalar negative log-likelihood value.
}
\description{
Negative log-likelihood under SER with summary stats (internal)
}
\keyword{internal}
