% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_priors.R
\name{neg_log_lik_var}
\alias{neg_log_lik_var}
\title{Negative log-likelihood parameterization: variance only (internal)}
\usage{
neg_log_lik_var(sigma_0_2, mu_0, X, y, sigma_2, prior_inclusion_weights)
}
\arguments{
\item{sigma_0_2}{Log prior variance (scalar or length-p).}

\item{mu_0}{Length-p prior mean.}

\item{X}{Design matrix (n x p).}

\item{y}{Working residual vector (length n).}

\item{sigma_2}{Residual variance.}

\item{prior_inclusion_weights}{Length-p prior inclusion weights.}
}
\value{
Scalar negative log-likelihood value.
}
\description{
Negative log-likelihood parameterization: variance only (internal)
}
\keyword{internal}
