% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_priors_ss.R
\name{neg_log_lik_var_ss}
\alias{neg_log_lik_var_ss}
\title{Negative log-likelihood parameterization (summary stats): variance only (internal)}
\usage{
neg_log_lik_var_ss(
  sigma_0_2,
  mu_0,
  Xty,
  dXtX,
  sigma_2,
  prior_inclusion_weights
)
}
\arguments{
\item{sigma_0_2}{Log prior variance (scalar or length-p).}

\item{mu_0}{Length-p prior mean.}

\item{Xty}{p-vector X^T y.}

\item{dXtX}{p-vector = diag(X^T X).}

\item{sigma_2}{Residual variance.}

\item{prior_inclusion_weights}{Length-p prior inclusion weights.}
}
\value{
Scalar negative log-likelihood value.
}
\description{
Negative log-likelihood parameterization (summary stats): variance only (internal)
}
\keyword{internal}
