% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SER.R
\name{SER}
\alias{SER}
\title{Single Effect Regression (SER)}
\usage{
SER(X, y, sigma_2, priors)
}
\arguments{
\item{X}{Observed data - matrix}

\item{y}{Observed response - vector}

\item{sigma_2}{Residual variance - scalar}

\item{priors}{List of prior parameters}
}
\value{
list(mu_1, sigma_1_2, BF), where:
mu_1 is the posterior effect mean vector
sigma_1_2 is the posterior effect variance vector
BF is the Bayesian factor vector
}
\description{
Single Effect Regression (SER)
}
