% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERSS_ss.R
\name{SER_ERSS_ss}
\alias{SER_ERSS_ss}
\title{SER Posterior Expectation of ||y - X b_l||^2 (summary stats)}
\usage{
SER_ERSS_ss(dXtX, Xty, yty, b_hat, b_2_hat)
}
\arguments{
\item{dXtX}{p-vector = diag(X^T X) (standardized scale).}

\item{Xty}{p-vector X^T y (standardized scale).}

\item{yty}{Scalar y^T y (unused here but kept for symmetry).}

\item{b_hat}{Posterior mean vector for a single effect (length p).}

\item{b_2_hat}{Posterior second moment vector for a single effect (length p).}
}
\value{
Numeric scalar \eqn{E[||y - X b_l||^2]} using summary statistics.
}
\description{
SER Posterior Expectation of ||y - X b_l||^2 (summary stats)
}
