% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{standardize_X}
\alias{standardize_X}
\title{Standardize X data}
\usage{
standardize_X(X)
}
\arguments{
\item{X}{(n x p) Data matrix}
}
\value{
X_scaled -- a centered and scaled (n x p) matrix
}
\description{
Standardize X data
}
